Fix a test issue with IPv6 localhost interfaces.
This fixes a regression introduced by 24c4bc0dd4b854291073e72eb2505f0207b9719b. It could occasionally cause the tests to fail when run on a machine with an IPv6 localhost interface. As a fix I added a wrapper for the new Go 1.15 `LookupIP()` method, and created a partially-functional backport for Go 1.14. This should be easy to delete in the future.

Signed-off-by: Matt Moyer <moyerm@vmware.com>