De-dup markdown docs by deleting them in site/ tree
I'm not sure if these docs are used anywhere in our website, but I don't think
that they are. I'm assuming someone or something will yell if these should not
be deleted. These docs also live at the root of the repo, and the duplicate
versions are already drifting out of sync from one another.

Signed-off-by: Andrew Keesler <akeesler@vmware.com>