Revert "Temporarily disable max inflight checks for mutating requests"
This reverts commit 4a28d1f80083fb7a46180c102e65c1e5a6f977d6.

This commit was originally made to fix a bug that caused TokenCredentialRequest
to become slow when the server was idle for an extended period of time. This was
to address a Kubernetes issue that was fixed in 1.19.5 and onward. We are now
running with Kubernetes 1.20, so we should be able to pick up this fix.