Add CLI caching of Concierge temporary credentials.
This adds a second cache file under `~/.config/pinniped/cluster-credentials.yaml`.
This file holds cached credentials returned by the Concierge's TokenCredentialRequest API.
The cache key is computed as an opaque hash over the input token as well as the destination endpoint (URL, authenticator name/type, and CA).
The logic is heavily borrowed from the existing "filesession" cache we have for OIDC tokens.
This cache is a bit simpler but has similar semantics about pruning expired or old, unused entries.

The overall goal of this cache is to reduce the median latency of `kubectl` commands that use the Concierge.

Signed-off-by: Matt Moyer <moyerm@vmware.com>