Change access token storage TTL to match refresh.
This is a partial cherry-pick of 5240f5e84ad3a984d3331adab3a6383ae09b7450. The token expirations are unchanged, but the garbage collection lifetime is now matched to avoid garbage collection breaking the refresh flow.

This is a backport to fix https://github.com/vmware-tanzu/pinniped/issues/601 on the v0.4.x release line.

Signed-off-by: Matt Moyer <moyerm@vmware.com>