Use a custom type for our static CLI client.
Before this change, we used the `fosite.DefaultOpenIDConnectClient{}` struct, which implements the  `fosite.Client` and `fosite.OpenIDConnectClient` interfaces. For a future change, we also need to implement some additional optional interfaces, so we can no longer use the provided default types. Instead, we now use a custom `staticclientt.PinnipedCLI{}` struct, which implements all the requisite interfaces and can be extended to handle the new functionality (in a future change).

The only user-observable change is that the storage format has been simplified. Instead of a full JSON serialization of the client, we now store only the client ID (`{"id": "pinniped-cli"}`). This is backwards compatible when existing refresh sessions or other stored values are read by the new code. There will be a brief (~60s) period of unavailability during upgrades due to older Supervisor pods being unable to correctly deserialize the new JSON.

Signed-off-by: Matt Moyer <moyerm@vmware.com>